app/requirements.txt
numpy
scipy
librosa
spleeter
demucs
whisper-openai
pysrt
pydub
aeneas
moviepy
ffmpeg-python